+++ pkg/client/client_v0040_test.go
@@ -4,7 +4,6 @@
 package client
 
 import (
-	"context"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -39,12 +38,6 @@ var _ = Describe("Client", func() {
 			cl, err = NewClient(cfg)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(cl).NotTo(BeNil())
-
-			go cl.Start(context.TODO())
-
-			DeferCleanup(func() {
-				cl.Stop()
-			})
 		})
 
 		Context("Get", func() {
@@ -76,12 +69,6 @@ var _ = Describe("Client", func() {
 			cl, err = NewClient(cfg)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(cl).NotTo(BeNil())
-
-			go cl.Start(context.TODO())
-
-			DeferCleanup(func() {
-				cl.Stop()
-			})
 		})
 
 		BeforeEach(func() {
@@ -220,12 +207,6 @@ var _ = Describe("Client", func() {
 			cl, err = NewClient(cfg)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(cl).NotTo(BeNil())
-
-			go cl.Start(context.TODO())
-
-			DeferCleanup(func() {
-				cl.Stop()
-			})
 		})
 
 		Context("Delete", func() {
@@ -295,12 +276,6 @@ var _ = Describe("Client", func() {
 			cl, err = NewClient(cfg)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(cl).NotTo(BeNil())
-
-			go cl.Start(context.TODO())
-
-			DeferCleanup(func() {
-				cl.Stop()
-			})
 		})
 
 		Context("Get", func() {